requirements.txt
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
mouseinfo==0.1.3
pillow==11.0.0
pip==24.3.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1-modules==0.4.1
pyautogui==0.9.54
pydantic==2.10.4
pydantic-core==2.27.2
pygetwindow==0.0.9
pymsgbox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
pyrect==0.2.0
pyscreeze==1.0.1
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
requests==2.32.3
rsa==4.9
tqdm==4.67.1
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
